Implement basic indicator plugin.
This provides the basic plugin with a working indicator. We expose
a way to access the docklet menu and set that directly as the indicator
menu. It's functional, but doesn't update to reflect program state as
it changes, has some issues with certain stock menu items and crashes
when you try to quit.